+++ pyfreebilling/settings.py
@@ -126,6 +126,7 @@ MIDDLEWARE_CLASSES = (
     'django.contrib.sessions.middleware.SessionMiddleware',
     'django.middleware.csrf.CsrfViewMiddleware',
     'django.contrib.auth.middleware.AuthenticationMiddleware',
+    'request.middleware.RequestMiddleware',
     'django.contrib.messages.middleware.MessageMiddleware',
 #    'djangosecure.middleware.SecurityMiddleware',
     'axes.middleware.FailedLoginMiddleware',
@@ -187,6 +188,8 @@ INSTALLED_APPS = (
     'did',
     'simple_import',
     'sysmon',
+    'request',
+    'model_report',
     #'massadmin',
 )